fix for first-time openid login
Old situation: If logged in in LXA, the OpenID login (e.g. on SF) worked.
If not logged in, it went to the LXA login and logged in fine there but SF gave an error.
Next time it worked (because now you were logged in in LXA).

New situation: This issue is fixed, it also works directly. I.e. you are not logged in
in LXA, you start the OpenID login (e.g. on SF), it goes to the LXA login, you loggin
and it returns to SF which also successfully loggs in.